Move filters to filter factory
