+++ blackboard/src/main/scala/asso/knowledgeSources/FilterFactory.scala
@@ -1,5 +1,7 @@
 package asso.knowledgeSources
 
+import asso.message.{Eof, Value}
+
 object FilterFactory {
 
   private def isPrime(n: Long): Boolean = {
@@ -29,4 +31,63 @@ object FilterFactory {
   def buildJoinMultiplesFilter(): JoinConditionFilter[Long] = {
     return new JoinConditionFilter[Long](isNotMultiple)
   }
+}
+
+case class ConditionFilter[T](val condition: (T) => Boolean) extends KnowledgeSource[T] {
+
+  override def execute() {
+    if (haveMessages()){
+      val message = messagesQueue1.dequeue()
+      message.setTopic(nextTopic)
+      message match {
+        case Value(value1, _) => {
+          if (condition(value1)) {
+            // advance object stage and put it back in the blackboard
+            blackboard.addToQueue(message)
+          }
+        }
+        case Eof(_) => {
+          blackboard.addToQueue(message);
+          println("Condition Filter Finished")
+          receivedEof = true;
+        }
+      }
+    }
+
+  }
+}
+
+
+case class JoinConditionFilter[T](val condition: (T) => Boolean) extends KnowledgeSource[T] {
+
+  // needs to receive 2 EOF to terminate
+  var receivedFirstEof = false;
+
+  override def execute() {
+    if (haveMessages()){
+      val message = messagesQueue1.dequeue()
+      message.setTopic(nextTopic)
+      message match {
+        case Value(value1, _) => {
+          if (condition(value1)) {
+            blackboard.addToQueue(message)
+          }
+        }
+        case Eof(_) => {
+          if (receivedFirstEof) {
+            receivedEof = true
+            blackboard.addToQueue(message)
+            println("Condition Filter Finished")
+          }
+          receivedFirstEof = true
+        }
+      }
+    }
+  }
+
+  override def setTopic(topic: Int): Unit = {
+    blackboard.addObserver(this, topic)
+  }
+
+
 }
